+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/AudioChannel.java
@@ -3,6 +3,7 @@ package de.maxhenkel.voicechat.voice.client;
 import de.maxhenkel.voicechat.Voicechat;
 import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.api.opus.OpusDecoder;
+import de.maxhenkel.voicechat.debug.VoicechatUncaughtExceptionHandler;
 import de.maxhenkel.voicechat.integration.freecam.FreecamUtil;
 import de.maxhenkel.voicechat.plugins.PluginManager;
 import de.maxhenkel.voicechat.plugins.impl.opus.OpusManager;
@@ -49,6 +50,7 @@ public class AudioChannel extends Thread {
         this.minecraft = Minecraft.getMinecraft();
         setDaemon(true);
         setName("AudioChannelThread-" + uuid.toString());
+        setUncaughtExceptionHandler(new VoicechatUncaughtExceptionHandler());
         Voicechat.LOGGER.info("Creating audio channel for " + uuid);
     }