+++ common/src/main/java/de/maxhenkel/voicechat/voice/server/Server.java
@@ -5,6 +5,7 @@ import de.maxhenkel.voicechat.api.RawUdpPacket;
 import de.maxhenkel.voicechat.api.VoicechatSocket;
 import de.maxhenkel.voicechat.api.events.SoundPacketEvent;
 import de.maxhenkel.voicechat.debug.CooldownTimer;
+import de.maxhenkel.voicechat.debug.VoicechatUncaughtExceptionHandler;
 import de.maxhenkel.voicechat.intercompatibility.CommonCompatibilityManager;
 import de.maxhenkel.voicechat.permission.PermissionManager;
 import de.maxhenkel.voicechat.plugins.PluginManager;
@@ -65,6 +66,7 @@ public class Server extends Thread {
         categoryManager = new ServerCategoryManager(this);
         setDaemon(true);
         setName("VoiceChatServerThread");
+        setUncaughtExceptionHandler(new VoicechatUncaughtExceptionHandler());
         processThread = new ProcessThread();
         processThread.start();
     }
@@ -179,6 +181,7 @@ public class Server extends Thread {
             lastKeepAlive = 0L;
             setDaemon(true);
             setName("VoiceChatPacketProcessingThread");
+            setUncaughtExceptionHandler(new VoicechatUncaughtExceptionHandler());
         }
 
         @Override